fix popup panel resizing
